Convert VectorId from `u32` to `u64`

+++ src/api/vectordb/vectors/dtos.rs
@@ -4,13 +4,13 @@ use crate::models::rpc::Vector;
 
 #[derive(Deserialize)]
 pub(crate) struct CreateVectorDto {
-    pub id: u32,
+    pub id: u64,
     pub values: Vec<f32>,
 }
 
 #[derive(Serialize)]
 pub(crate) struct CreateVectorResponseDto {
-    pub id: u32,
+    pub id: u64,
     pub values: Vec<f32>,
     // pub created_at: String
 }
@@ -22,7 +22,7 @@ pub(crate) struct UpdateVectorDto {
 
 #[derive(Serialize)]
 pub(crate) struct UpdateVectorResponseDto {
-    pub id: u32,
+    pub id: u64,
     pub values: Vec<f32>,
     // pub created_at: String
 }
@@ -30,12 +30,12 @@ pub(crate) struct UpdateVectorResponseDto {
 #[derive(Deserialize)]
 pub(crate) struct FindSimilarVectorsDto {
     pub vector: Vec<f32>,
-    pub k: u32,
+    pub k: u64,
 }
 
 #[derive(Serialize)]
 pub(crate) struct SimilarVector {
-    pub id: u32,
+    pub id: u64,
     pub score: f32,
 }
