+++ src/models/serializer/prob/tests.rs
@@ -114,7 +114,7 @@ fn get_cache(bufmans: Arc<BufferManagerFactory>, prop_file: Arc<RwLock<File>>) -
     Arc::new(ProbCache::new(1000, bufmans, prop_file))
 }
 
-fn create_prob_node(id: u32, prop_file: &RwLock<File>) -> ProbNode {
+fn create_prob_node(id: u64, prop_file: &RwLock<File>) -> ProbNode {
     let id = VectorId(id);
     let value = Arc::new(Storage::UnsignedByte {
         mag: 10,
